flyte_demo/requirements.txt
dataclasses_json
flytekit>=1.1.0
flytekitplugins-bigquery
flytekitplugins-deck-standard
flytekitplugins-pandera
flytekitplugins-spark
ipykernel
joblib
modin
palmerpenguins
pandas
pyspark
scikit-learn
torch